added tabs and charts skeleton
